Rebase reconstruction to native cluster and track classes + related changes

+++ rec/include/NA6PTrackerCA.h
@@ -29,10 +29,10 @@ struct TrackletCandidate {
   int startingLayer;
   int firstClusterIndex;
   int secondClusterIndex;
-  double tanL;
-  double phi;
-  double pxpz;
-  double pypz;
+  float tanL;
+  float phi;
+  float pxpz;
+  float pypz;
 };
 
 struct CellCandidate {
@@ -59,7 +59,7 @@ struct TrackFitted {
   std::vector<int> cluIDs;
   NA6PTrack trackFitFast;
   //    genfit::Track trackFit;
-  double chi2ndf;
+  float chi2ndf;
 };
 
 enum class ExtendDirection { kInward,
@@ -79,15 +79,15 @@ class NA6PTrackerCA
   void setMaxNumberOfSharedClusters(int n) { mMaxSharedClusters = n; }
   void setNumberOfIterations(int nIter);
   void setIterationParams(int iter,
-                          double maxDeltaThetaTracklets,
-                          double maxDeltaPhiTracklets,
-                          double maxDeltaTanLCells,
-                          double maxDeltaPhiCells,
-                          double maxDeltaPxPzCells,
-                          double maxDeltaPyPzCells,
-                          double maxChi2TrClCells,
-                          double maxChi2ndfCells,
-                          double maxChi2ndfTracks,
+                          float maxDeltaThetaTracklets,
+                          float maxDeltaPhiTracklets,
+                          float maxDeltaTanLCells,
+                          float maxDeltaPhiCells,
+                          float maxDeltaPxPzCells,
+                          float maxDeltaPyPzCells,
+                          float maxChi2TrClCells,
+                          float maxChi2ndfCells,
+                          float maxChi2ndfTracks,
                           int minNClusTracks);
   void configureFromRecoParam(const std::string& filename = "");
   void setVerbosity(bool opt = true) { mVerbose = opt; }
@@ -95,112 +95,112 @@ class NA6PTrackerCA
   void printConfiguration() const;
   int getNIterations() const { return mNIterationsCA; }
   bool loadGeometry(const char* filename, const char* geoname = "NA6P");
-  
+
   template<typename ClusterType>
   void findTracks(std::vector<ClusterType>& cluArr, TVector3 primVert);
-  
+
   std::vector<NA6PTrack> getTracks();
 
 protected:
   // methods used in tracking
-  template<typename ClusterType>
-  void sortClustersByLayerAndEta(std::vector<ClusterType>& cluArr,
-				 std::vector<int>& firstIndex,
-				 std::vector<int>& lastIndex);
-  void sortTrackletsByLayerAndIndex(std::vector<TrackletCandidate>& tracklets,
-				    std::vector<int>& firstIndex,
-				    std::vector<int>& lastIndex);
-  void sortCellsByLayerAndIndex(std::vector<CellCandidate>& cells,
-				std::vector<int>& firstIndex,
-				std::vector<int>& lastIndex);
-  template<typename ClusterType>
-  void computeLayerTracklets(const std::vector<ClusterType>& cluArr,
-			     const std::vector<int>& firstIndex,
-			     const std::vector<int>& lastIndex,
-			     std::vector<TrackletCandidate>& tracklets,
-			     double deltaThetaMax,
-			     double deltaPhiMax);
-  template<typename ClusterType>
-  void computeLayerCells(const std::vector<TrackletCandidate>& tracklets,
-			 const std::vector<int>& firstIndex,
-			 const std::vector<int>& lastIndex,
-			 const std::vector<ClusterType>& cluArr,
-			 std::vector<CellCandidate>& cells,
-			 double deltaTanLMax,
-			 double deltaPhiMax,
-			 double deltaPxPzMax,
-			 double deltaPyPzMax,
-			 double maxChi2TrClu,
-			 double maxChi2NDF);
-  template<typename ClusterType>
-  double computeTrackToClusterChi2(const NA6PTrack& track,
-				   const ClusterType& clu);
-  template<typename ClusterType>
-  bool fitTrackPointsFast(const std::vector<int>& cluIDs,
-			  const std::vector<ClusterType>& cluArr,
-			  NA6PTrack& fitTrack,
-			  double maxChi2TrClu,
-			  double maxChi2NDF);
-  template<typename ClusterType>
-  void findCellsNeighbours(const std::vector<CellCandidate>& cells,
-			   const std::vector<int>& firstIndex,
-			   const std::vector<int>& lastIndex,
-			   std::vector<std::pair<int,int>>& cneigh,
-			   const std::vector<ClusterType>& cluArr,
-			   double maxChi2TrClu);
-  template<typename ClusterType>
-  std::vector<TrackCandidate> prolongSeed(const TrackCandidate& seed,
-					  const std::vector<CellCandidate>& cells,
-					  const std::vector<int>& firstIndex,
-					  const std::vector<int>& lastIndex,
-					  const std::vector<ClusterType>& cluArr,
-					  double maxChi2TrClu,
-					  ExtendDirection dir);
-  template<typename ClusterType>
-  void findRoads(const std::vector<std::pair<int,int>>& cneigh,
-		 const std::vector<CellCandidate>& cells,
-		 const std::vector<int>& firstIndex,
-		 const std::vector<int>& lastIndex,
-		 const std::vector<TrackletCandidate>& tracklets,
-		 const std::vector<ClusterType>& cluArr,
-		 std::vector<TrackCandidate>& trackCands,
-		 double maxChi2TrClu);
-  template<typename ClusterType>
-  void fitAndSelectTracks(const std::vector<TrackCandidate>& trackCands,
-			  const std::vector<ClusterType>& cluArr,
-			  std::vector<TrackFitted>& tracks,
-			  double maxChi2TrClu,
-			  int minNClu,
-			  double maxChi2NDF);
-  template<typename T, typename ClusterType>
-  void printStats(const std::vector<T>& candidates,
-		  const std::vector<ClusterType>& cluArr,
-		  const std::vector<CellCandidate>& cells,
-		  const std::string& label,
-		  int requiredClus = -1);
+ template <typename ClusterType>
+ void sortClustersByLayerAndEta(std::vector<ClusterType>& cluArr,
+                                std::vector<int>& firstIndex,
+                                std::vector<int>& lastIndex);
+ void sortTrackletsByLayerAndIndex(std::vector<TrackletCandidate>& tracklets,
+                                   std::vector<int>& firstIndex,
+                                   std::vector<int>& lastIndex);
+ void sortCellsByLayerAndIndex(std::vector<CellCandidate>& cells,
+                               std::vector<int>& firstIndex,
+                               std::vector<int>& lastIndex);
+ template <typename ClusterType>
+ void computeLayerTracklets(const std::vector<ClusterType>& cluArr,
+                            const std::vector<int>& firstIndex,
+                            const std::vector<int>& lastIndex,
+                            std::vector<TrackletCandidate>& tracklets,
+                            float deltaThetaMax,
+                            float deltaPhiMax);
+ template <typename ClusterType>
+ void computeLayerCells(const std::vector<TrackletCandidate>& tracklets,
+                        const std::vector<int>& firstIndex,
+                        const std::vector<int>& lastIndex,
+                        const std::vector<ClusterType>& cluArr,
+                        std::vector<CellCandidate>& cells,
+                        float deltaTanLMax,
+                        float deltaPhiMax,
+                        float deltaPxPzMax,
+                        float deltaPyPzMax,
+                        float maxChi2TrClu,
+                        float maxChi2NDF);
+ template <typename ClusterType>
+ float computeTrackToClusterChi2(const NA6PTrack& track,
+                                 const ClusterType& clu);
+ template <typename ClusterType>
+ bool fitTrackPointsFast(const std::vector<int>& cluIDs,
+                         const std::vector<ClusterType>& cluArr,
+                         NA6PTrack& fitTrack,
+                         float maxChi2TrClu,
+                         float maxChi2NDF);
+ template <typename ClusterType>
+ void findCellsNeighbours(const std::vector<CellCandidate>& cells,
+                          const std::vector<int>& firstIndex,
+                          const std::vector<int>& lastIndex,
+                          std::vector<std::pair<int, int>>& cneigh,
+                          const std::vector<ClusterType>& cluArr,
+                          float maxChi2TrClu);
+ template <typename ClusterType>
+ std::vector<TrackCandidate> prolongSeed(const TrackCandidate& seed,
+                                         const std::vector<CellCandidate>& cells,
+                                         const std::vector<int>& firstIndex,
+                                         const std::vector<int>& lastIndex,
+                                         const std::vector<ClusterType>& cluArr,
+                                         float maxChi2TrClu,
+                                         ExtendDirection dir);
+ template <typename ClusterType>
+ void findRoads(const std::vector<std::pair<int, int>>& cneigh,
+                const std::vector<CellCandidate>& cells,
+                const std::vector<int>& firstIndex,
+                const std::vector<int>& lastIndex,
+                const std::vector<TrackletCandidate>& tracklets,
+                const std::vector<ClusterType>& cluArr,
+                std::vector<TrackCandidate>& trackCands,
+                float maxChi2TrClu);
+ template <typename ClusterType>
+ void fitAndSelectTracks(const std::vector<TrackCandidate>& trackCands,
+                         const std::vector<ClusterType>& cluArr,
+                         std::vector<TrackFitted>& tracks,
+                         float maxChi2TrClu,
+                         int minNClu,
+                         float maxChi2NDF);
+ template <typename T, typename ClusterType>
+ void printStats(const std::vector<T>& candidates,
+                 const std::vector<ClusterType>& cluArr,
+                 const std::vector<CellCandidate>& cells,
+                 const std::string& label,
+                 int requiredClus = -1);
 
- private:
-  int mNLayers = 5;
-  int mLayerStart = 0;
-  double mPrimVertPos[3] = {};
-  NA6PFastTrackFitter* mTrackFitter = nullptr;
-  std::vector<bool> mIsClusterUsed = {};
-  std::vector<TrackFitted> mFinalTracks = {};
-  int mMaxSharedClusters = 0;
-  bool mVerbose = false;
-  int mNIterationsCA = 2;
-  double mMaxDeltaThetaTrackletsCA[kMaxIterationsCA] = {0.04, 0.1, 0.15, 0.3, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxDeltaPhiTrackletsCA[kMaxIterationsCA] = {0.1, 0.2, 0.25, 0.5, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxDeltaTanLCellsCA[kMaxIterationsCA] = {4., 9., 18., 40., 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxDeltaPhiCellsCA[kMaxIterationsCA] = {0.4, 0.6, 1.2, 2.4, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxDeltaPxPzCellsCA[kMaxIterationsCA] = {0.02, 0.05, 0.1, 0.2, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxDeltaPyPzCellsCA[kMaxIterationsCA] = {2e-3, 7.5e-3, 1.5e-2, 3.e-2, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxChi2TrClCellsCA[kMaxIterationsCA] = {100., 500., 999., 999.0, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxChi2ndfCellsCA[kMaxIterationsCA] = {100., 500., 999.0, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  double mMaxChi2ndfTracksCA[kMaxIterationsCA] = {100., 500., 999.0, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0, 999.0};
-  int    mMinNClusTracksCA[kMaxIterationsCA] = {5, 3, 0, 0, 0, 0, 0, 0, 0, 0};
+private:
+ int mNLayers = 5;
+ int mLayerStart = 0;
+ float mPrimVertPos[3] = {};
+ NA6PFastTrackFitter* mTrackFitter = nullptr;
+ std::vector<bool> mIsClusterUsed = {};
+ std::vector<TrackFitted> mFinalTracks = {};
+ int mMaxSharedClusters = 0;
+ bool mVerbose = false;
+ int mNIterationsCA = 2;
+ float mMaxDeltaThetaTrackletsCA[kMaxIterationsCA] = {0.04f, 0.1f, 0.15f, 0.3f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxDeltaPhiTrackletsCA[kMaxIterationsCA] = {0.1f, 0.2f, 0.25f, 0.5f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxDeltaTanLCellsCA[kMaxIterationsCA] = {4.f, 9.f, 18.f, 40.f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxDeltaPhiCellsCA[kMaxIterationsCA] = {0.4f, 0.6f, 1.2f, 2.4f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxDeltaPxPzCellsCA[kMaxIterationsCA] = {0.02f, 0.05f, 0.1f, 0.2f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxDeltaPyPzCellsCA[kMaxIterationsCA] = {2e-3f, 7.5e-3f, 1.5e-2f, 3.e-2f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxChi2TrClCellsCA[kMaxIterationsCA] = {100.f, 500.f, 999.f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxChi2ndfCellsCA[kMaxIterationsCA] = {100.f, 500.f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ float mMaxChi2ndfTracksCA[kMaxIterationsCA] = {100.f, 500.f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f, 999.0f};
+ int mMinNClusTracksCA[kMaxIterationsCA] = {5, 3, 0, 0, 0, 0, 0, 0, 0, 0};
 
-  ClassDefNV(NA6PTrackerCA, 1);
+ ClassDefNV(NA6PTrackerCA, 1);
 };
 
 #endif
